Rewrite with expect/actual
When using this multiplatform library in a java runtime, the interfaces and classes should now point to the standard java otel implementation classes, no longer requiring a wrapper object.
The implementation for js is still pending. Given that we match the java interfaces so closely, we will probably still have to use wrappers for js in many (all?) cases.
Api-Events is also still missing completely, api-logs has been scrapped.